chore: update rust example dependencies and move metrics into the 'spin-basic' example
