+++ rust/Cargo.toml
@@ -11,19 +11,19 @@ spin-sdk = { workspace = true }
 tracing = { workspace = true }
 tracing-opentelemetry = { workspace = true }
 tracing-subscriber = { workspace = true }
-wit-bindgen = "0.30.0"
+wit-bindgen = "0.47.0"
 
 [workspace]
 members = ["examples/*"]
 
 [workspace.dependencies]
 anyhow = "1"
-opentelemetry = { version = "0.30.0" }
-opentelemetry_sdk = { version = "0.30.0", features = ["experimental_metrics_custom_reader"] }
-spin-sdk = "3.1.0"
+opentelemetry = { version = "0.31.0" }
+opentelemetry_sdk = { version = "0.31.0", features = ["experimental_metrics_custom_reader"] }
+spin-sdk = "5.1.1"
 tracing = "0.1.41"
-tracing-opentelemetry = "0.31.0"
-tracing-subscriber = "0.3.19"
+tracing-opentelemetry = "0.32.0"
+tracing-subscriber = "0.3.20"
 
 [features]
 # This enables OTel internal logging.
