--- rust/examples/spin-metrics/.gitignore
@@ -1,2 +1,0 @@
-target/
-.spin/
