--- rust/examples/spin-metrics/Cargo.toml
@@ -1,17 +1,0 @@
-[package]
-name = "spin-metrics"
-authors = ["Andrew Steurer <94206073+asteurer@users.noreply.github.com>"]
-description = ""
-version = "0.1.0"
-rust-version = "1.78"
-edition = "2021"
-
-[lib]
-crate-type = ["cdylib"]
-
-[dependencies]
-anyhow = { workspace = true }
-opentelemetry = { workspace = true }
-opentelemetry_sdk = { workspace = true }
-opentelemetry-wasi = { path = "../../" }
-spin-sdk = { workspace = true }
