--- rust/examples/spin-metrics/spin.toml
@@ -1,18 +1,0 @@
-spin_manifest_version = 2
-
-[application]
-name = "spin-metrics"
-version = "0.1.0"
-authors = ["Andrew Steurer <94206073+asteurer@users.noreply.github.com>"]
-description = "A Spin application demonstrating the use of OpenTelemetry metrics"
-
-[[trigger.http]]
-route = "/..."
-component = "spin-metrics"
-
-[component.spin-metrics]
-source = "../../target/wasm32-wasip2/release/spin_metrics.wasm"
-allowed_outbound_hosts = []
-[component.spin-metrics.build]
-command = "cargo build --target wasm32-wasip2 --release"
-watch = ["src/**/*.rs", "Cargo.toml"]
